# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Jarvis Brain API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Vault {
  id: number;
  name: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  icon?: string | null;
  /** @nullable */
  color?: string | null;
  createdAt: string;
  updatedAt: string;
}

export interface VaultInput {
  /** @minLength 1 */
  name: string;
  description?: string;
  icon?: string;
  color?: string;
}

export interface Note {
  id: number;
  vaultId: number;
  title: string;
  content: string;
  /** @nullable */
  excerpt?: string | null;
  tags?: string[];
  createdAt: string;
  updatedAt: string;
}

export interface NoteInput {
  vaultId: number;
  /** @minLength 1 */
  title: string;
  content: string;
  excerpt?: string;
  tags?: string[];
}

export interface Tag {
  id: number;
  name: string;
  /** @nullable */
  color?: string | null;
  createdAt?: string;
}

export interface TagInput {
  /** @minLength 1 */
  name: string;
  color?: string;
}

export interface NoteLink {
  id: number;
  sourceNoteId: number;
  targetNoteId: number;
  /** @nullable */
  relation?: string | null;
  createdAt?: string;
}

export interface NoteLinkInput {
  targetNoteId: number;
  relation?: string;
}

export type GraphNodesItem = {
  id: number;
  title: string;
  x: number;
  y: number;
  /** @nullable */
  color?: string | null;
};

export type GraphEdgesItem = {
  source: number;
  target: number;
  /** @nullable */
  relation?: string | null;
};

export interface Graph {
  nodes: GraphNodesItem[];
  edges: GraphEdgesItem[];
}

export interface Stats {
  totalNotes: number;
  totalTags: number;
  totalVaults: number;
  totalLinks: number;
  recentNotes: Note[];
}

export type GetVaultNotesParams = {
tag?: string;
q?: string;
};

export type GetGraphParams = {
vaultId?: number;
};

export type SearchNotesParams = {
q: string;
vaultId?: number;
};