# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Jarvis Brain API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: notes
  - name: tags
  - name: vaults
  - name: graph
  - name: search
  - name: stats
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /vaults:
    get:
      operationId: listVaults
      tags: [vaults]
      summary: List all vaults
      responses:
        "200":
          description: List of vaults
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Vault"
    post:
      operationId: createVault
      tags: [vaults]
      summary: Create a new vault
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultInput"
      responses:
        "201":
          description: Vault created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"

  /vaults/{vaultId}:
    get:
      operationId: getVault
      tags: [vaults]
      summary: Get a vault
      parameters:
        - name: vaultId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Vault
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
    patch:
      operationId: updateVault
      tags: [vaults]
      summary: Update a vault
      parameters:
        - name: vaultId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VaultInput"
      responses:
        "200":
          description: Vault updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vault"
    delete:
      operationId: deleteVault
      tags: [vaults]
      summary: Delete a vault
      parameters:
        - name: vaultId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /vaults/{vaultId}/notes:
    get:
      operationId: getVaultNotes
      tags: [notes]
      summary: List notes in vault
      parameters:
        - name: vaultId
          in: path
          required: true
          schema:
            type: integer
        - name: tag
          in: query
          schema:
            type: string
        - name: q
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Notes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

  /notes:
    post:
      operationId: createNote
      tags: [notes]
      summary: Create a note
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "201":
          description: Note created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"

  /notes/{noteId}:
    get:
      operationId: getNote
      tags: [notes]
      summary: Get a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Note
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    patch:
      operationId: updateNote
      tags: [notes]
      summary: Update a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteInput"
      responses:
        "200":
          description: Note updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Note"
    delete:
      operationId: deleteNote
      tags: [notes]
      summary: Delete a note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /notes/{noteId}/tags:
    get:
      operationId: getNoteTags
      tags: [tags]
      summary: Get note tags
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Tags
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Tag"
    post:
      operationId: addNoteTag
      tags: [tags]
      summary: Add tag to note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TagInput"
      responses:
        "201":
          description: Tag added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tag"

  /notes/{noteId}/tags/{tagId}:
    delete:
      operationId: removeNoteTag
      tags: [tags]
      summary: Remove tag from note
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
        - name: tagId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed

  /notes/{noteId}/links:
    get:
      operationId: getNoteLinks
      tags: [graph]
      summary: Get note links
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Links
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NoteLink"
    post:
      operationId: createNoteLink
      tags: [graph]
      summary: Link notes
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NoteLinkInput"
      responses:
        "201":
          description: Link created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NoteLink"

  /notes/{noteId}/links/{linkId}:
    delete:
      operationId: deleteNoteLink
      tags: [graph]
      summary: Delete note link
      parameters:
        - name: noteId
          in: path
          required: true
          schema:
            type: integer
        - name: linkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  /graph:
    get:
      operationId: getGraph
      tags: [graph]
      summary: Get knowledge graph
      parameters:
        - name: vaultId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Graph
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Graph"

  /search:
    get:
      operationId: searchNotes
      tags: [search]
      summary: Search notes
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: vaultId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Note"

  /stats:
    get:
      operationId: getStats
      tags: [stats]
      summary: Get stats
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stats"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Vault:
      type: object
      required: [id, name, createdAt, updatedAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: ["string", "null"] }
        icon: { type: ["string", "null"] }
        color: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    VaultInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        description: { type: string }
        icon: { type: string }
        color: { type: string }

    Note:
      type: object
      required: [id, vaultId, title, content, createdAt, updatedAt]
      properties:
        id: { type: integer }
        vaultId: { type: integer }
        title: { type: string }
        content: { type: string }
        excerpt: { type: ["string", "null"] }
        tags:
          type: array
          items: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }

    NoteInput:
      type: object
      required: [vaultId, title, content]
      properties:
        vaultId: { type: integer }
        title: { type: string, minLength: 1 }
        content: { type: string }
        excerpt: { type: string }
        tags:
          type: array
          items: { type: string }

    Tag:
      type: object
      required: [id, name]
      properties:
        id: { type: integer }
        name: { type: string }
        color: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }

    TagInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        color: { type: string }

    NoteLink:
      type: object
      required: [id, sourceNoteId, targetNoteId]
      properties:
        id: { type: integer }
        sourceNoteId: { type: integer }
        targetNoteId: { type: integer }
        relation: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }

    NoteLinkInput:
      type: object
      required: [targetNoteId]
      properties:
        targetNoteId: { type: integer }
        relation: { type: string }

    Graph:
      type: object
      required: [nodes, edges]
      properties:
        nodes:
          type: array
          items:
            type: object
            required: [id, title, x, y]
            properties:
              id: { type: integer }
              title: { type: string }
              x: { type: number }
              y: { type: number }
              color: { type: ["string", "null"] }
        edges:
          type: array
          items:
            type: object
            required: [source, target]
            properties:
              source: { type: integer }
              target: { type: integer }
              relation: { type: ["string", "null"] }

    Stats:
      type: object
      required: [totalNotes, totalTags, totalVaults, totalLinks, recentNotes]
      properties:
        totalNotes: { type: integer }
        totalTags: { type: integer }
        totalVaults: { type: integer }
        totalLinks: { type: integer }
        recentNotes:
          type: array
          items:
            $ref: "#/components/schemas/Note"